requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
google-generativeai==0.3.1
PyPDF2==3.0.1
Pillow==10.0.0
email-validator==2.0.0
gunicorn==21.2.0
python-magic==0.4.27